Refactor for AbstractMessage class for Encoder & Decoder.

+++ src/main/java/org/gamecenter/serializer/Decoder.java
@@ -33,7 +33,7 @@ public class Decoder extends Coder {
             byte[] bytes = ByteUtil.subBytes(msgBytes, HeaderFilter.TOTAL_HEADER_LENGTH, header.getMsgBodyLength());
 
 
-            int msgId = ByteUtil.getShort(header.getMessageId());
+            short msgId = ByteUtil.getShort(header.getMessageId());
 
             Message message = loader.getMessageByMsgId(msgId);
